refactor time series types
