+++ packages/stable/src/types.ts
@@ -41,6 +41,8 @@ export * from './exports.gen';
 export * from './api/geospatial/types';
 export * from './api/streams/types';
 export * from './api/records/types';
+export * from './api/sessions/types';
+export * from './api/timeSeries/types';
 
 export interface Acl<ActionsType, ScopeType> {
   actions: ActionsType[];
@@ -282,11 +284,6 @@ export type FileAggregate = AggregateResponse;
 export type SequenceAggregate = AggregateResponse;
 
 /**
- * Response from timeseries aggregate endpoint
- */
-export type TimeseriesAggregate = AggregateResponse;
-
-/**
  * Query schema for asset aggregate endpoint
  */
 export interface AssetAggregateQuery {
@@ -304,16 +301,6 @@ export interface FileAggregateQuery {
    * Filter on files with strict matching.
    */
   filter?: FileFilterProps;
-}
-
-/**
- * Query schema for timeseries aggregate endpoint
- */
-export interface TimeseriesAggregateQuery {
-  /**
-   * Filter on timeseries with strict matching.
-   */
-  filter?: TimeseriesFilter;
 }
 
 export interface AssetAggregateResult {
@@ -1226,93 +1213,6 @@ export interface DoubleDatapoint extends DatapointInfo {
 
 export interface StringDatapoint extends DatapointInfo {
   value: string;
-}
-
-export interface Timeseries extends InternalId, CreatedAndLastUpdatedTime {
-  /**
-   * Externally supplied id of the time series
-   */
-  externalId?: CogniteExternalId;
-  /**
-   * The ID of an instance in Cognite Data Models.
-   */
-  instanceId?: CogniteInstanceId;
-  name?: TimeseriesName;
-  isString: TimeseriesIsString;
-  /**
-   * Additional metadata. String key -> String value.
-   */
-  metadata?: Metadata;
-  unit?: TimeseriesUnit;
-  /**
-   * Asset that this time series belongs to.
-   */
-  assetId?: CogniteInternalId;
-  dataSetId?: CogniteInternalId;
-  isStep: TimeseriesIsStep;
-  /**
-   * Description of the time series.
-   */
-  description: string;
-  /**
-   * Security categories required in order to access this time series.
-   */
-  securityCategories?: number[];
-  /**
-   * The physical unit of the time series (reference to unit catalog).
-   */
-  unitExternalId?: CogniteExternalId;
-}
-
-export interface ExternalTimeseries {
-  /**
-   * Externally provided id for the time series (optional but recommended)
-   */
-  externalId?: CogniteExternalId;
-  /**
-   * Set a value for legacyName to allow applications using API v0.3, v04, v05, and v0.6 to access this time series. The legacy name is the human-readable name for the time series and is mapped to the name field used in API versions 0.3-0.6. The legacyName field value must be unique, and setting this value to an already existing value will return an error. We recommend that you set this field to the same value as externalId.
-   */
-  legacyName?: string;
-  /**
-   * Human readable name of time series
-   */
-  name?: string;
-  /**
-   * Whether the time series is string valued or not.
-   */
-  isString?: boolean;
-  /**
-   * Additional metadata. String key -> String value.
-   */
-  metadata?: Metadata;
-  /**
-   * The physical unit of the time series.
-   */
-  unit?: string;
-  /**
-   * Asset that this time series belongs to.
-   */
-  assetId?: CogniteInternalId;
-  /**
-   * DataSet that this time series related with.
-   */
-  dataSetId?: CogniteInternalId;
-  /**
-   * Whether the time series is a step series or not.
-   */
-  isStep?: boolean;
-  /**
-   * Description of the time series.
-   */
-  description?: string;
-  /**
-   * Security categories required in order to access this time series."
-   */
-  securityCategories?: number[];
-  /**
-   * The physical unit of the time series (reference to unit catalog).
-   */
-  unitExternalId?: CogniteExternalId;
 }
 
 export type FileGeoLocationType = 'Feature';
@@ -2338,162 +2238,6 @@ export const SortOrder = {
  */
 export type SortOrder = 'asc' | 'desc';
 
-export interface SyntheticDataError extends DatapointInfo {
-  error: string;
-}
-
-export type SyntheticDatapoint = SyntheticDataValue | SyntheticDataError;
-
-export interface SyntheticDataValue extends DatapointInfo {
-  value: number;
-}
-
-/**
- * A query for a synthetic time series
- */
-export interface SyntheticQuery extends Limit {
-  expression: string;
-  start?: string | Timestamp;
-  end?: string | Timestamp;
-}
-
-/**
- * Response of a synthetic time series query
- */
-export interface SyntheticQueryResponse {
-  isString?: TimeseriesIsString;
-  datapoints: SyntheticDatapoint[];
-}
-
-export type TimeseriesUpdateCommonProperies = {
-  externalId?: NullableSinglePatchString;
-  metadata?: MetadataPatch;
-  assetId?: NullableSinglePatchLong;
-  dataSetId?: NullableSinglePatchLong;
-};
-
-export type TimeseriesUpdateAssetCentricProperies =
-  TimeseriesUpdateCommonProperies & {
-    name?: NullableSinglePatchString;
-    unit?: NullableSinglePatchString;
-    description?: NullableSinglePatchString;
-    securityCategories?: ArrayPatchLong;
-    unitExternalId?: NullableSinglePatchString;
-  };
-
-export interface TimeSeriesPatch {
-  update: TimeseriesUpdateAssetCentricProperies;
-}
-
-export interface TimeSeriesPatchByInstanceId {
-  update: TimeseriesUpdateCommonProperies;
-}
-
-export interface TimeSeriesPatchByInstanceId {
-  update: {
-    externalId?: NullableSinglePatchString;
-    metadata?: MetadataPatch;
-    assetId?: NullableSinglePatchLong;
-    dataSetId?: NullableSinglePatchLong;
-  };
-}
-
-export interface TimeseriesSearch {
-  /**
-   * Prefix and fuzzy search on name.
-   */
-  name?: string;
-  /**
-   * Prefix and fuzzy search on description.
-   */
-  description?: string;
-  /**
-   * Search on name and description using wildcard search on each of the words (separated by spaces). Retrieves results where at least one word must match. Example: '*some* *other*'
-   */
-  query?: string;
-}
-
-export interface TimeseriesSearchFilter extends Limit {
-  filter?: TimeseriesFilter;
-  search?: TimeseriesSearch;
-}
-
-export type TimeSeriesUpdate =
-  | TimeSeriesUpdateById
-  | TimeSeriesUpdateByExternalId
-  | TimeSeriesUpdateByInstanceId;
-
-export interface TimeSeriesUpdateByExternalId
-  extends TimeSeriesPatch,
-    ExternalId {}
-
-export interface TimeSeriesUpdateById extends TimeSeriesPatch, InternalId {}
-
-export interface TimeSeriesUpdateByInstanceId
-  extends TimeSeriesPatchByInstanceId,
-    InstanceId {}
-
-export interface TimeseriesFilter extends CreatedAndLastUpdatedTimeFilter {
-  name?: TimeseriesName;
-  unit?: TimeseriesUnit;
-  isString?: TimeseriesIsString;
-  isStep?: TimeseriesIsStep;
-  metadata?: Metadata;
-  /**
-   * Get time series related to these assets. Takes [ 1 .. 100 ] unique items.
-   */
-  assetIds?: CogniteInternalId[];
-  /**
-   * Asset External IDs of related equipment that this time series relates to.
-   */
-  assetExternalIds?: CogniteExternalId[];
-  /**
-   * Only include timeseries that have a related asset in a tree rooted at any of these root assetIds.
-   */
-  rootAssetIds?: CogniteInternalId[];
-  /**
-   * Only include assets that reference these specific dataSet IDs
-   */
-  dataSetIds?: IdEither[];
-  /**
-   * Only include timeseries that are related to an asset in a subtree rooted at any of these assetIds.
-   * If the total size of the given subtrees exceeds 100,000 assets, an error will be returned.
-   */
-  assetSubtreeIds?: IdEither[];
-  externalIdPrefix?: ExternalIdPrefix;
-  /**
-   * The physical unit of the time series (reference to unit catalog).
-   */
-  unitExternalId?: CogniteExternalId;
-}
-
-export interface TimeseriesFilterQuery extends FilterQuery {
-  filter?: TimeseriesFilter;
-  partition?: Partition;
-}
-
-export type TimeseriesIdEither = InternalId | ExternalId;
-
-/**
- * Whether the time series is a step series or not.
- */
-export type TimeseriesIsStep = boolean;
-
-/**
- * Whether the time series is string valued or not.
- */
-export type TimeseriesIsString = boolean;
-
-/**
- * Name of time series
- */
-export type TimeseriesName = string;
-
-/**
- * The physical unit of the time series.
- */
-export type TimeseriesUnit = string;
-
 /**
  * A point in time, either a number or a Date object.
  * The Date is converted to a number when api calls are made.
@@ -2967,7 +2711,7 @@ export type SequencesSource = {
 
 export type TimeSeriesSource = {
   type: 'timeSeries';
-  filter?: ObjectOrString<TimeseriesFilter>;
+  filter?: ObjectOrString<import('./api/timeSeries/types').TimeseriesFilter>;
   mappings?: { [K in string]: string };
 };
 
@@ -3536,5 +3280,3 @@ export type {
   ViewCollectionResponseWithCursorResponse,
   ViewCreateCollection,
 } from './api/views/types.gen';
-
-export * from './api/sessions/types';
